pyproject.toml
[project]
name = "test-task"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic (>=2.12.4,<3.0.0)",
    "fastapi (>=0.121.0,<0.122.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "alembic (>=1.17.1,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "cryptography (>=46.0.3,<47.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
